feat: basic model
The basic model of the Court of the Blades:

SpecialAbility, SpecialFeature, XpTrigger are generic classes
to handle the Court of Blades basic model.

CoterieUpgrade represents a coterie upgrade.

CoterieUpgradeSource is anything which provides possible upgrades for
a coterie.

Coterie is the player group sharing resources and allegiance to
a single house.

Coterie Type represents various Coterie Playbooks.

HouseModel represents the noble house the coterie is serving.
* HouseModel is a CoterieUpgradeSource providing house specific
  upgrades for its servants.