feat: CoterieTestCase, NamedTestCase, TestCase
Created test kit for testing Coterie serialization in CoterieUnitTest,
using CoterieTestCase to store coterie test information, and the
CoterieTestInvocationContextProvider to generated the content for the
test.

Created TestCase and NamedTextCase for creation of test cases, and
their simple implementations SimpleTestCase and SimpleNamedTestCase.

Generated generic NamedTestInvocationContextProvider, and in the same
file the TestInvocationContextProvider.
* TODO: Move TestInvocationContextProvider to separate file.

ListConveter contains tools for converting generic lists to other
values, and extractor of a specific field ListEntryConveter.